requirements.txt
fastapi
uvicorn
paho-mqtt
websockets
mysql-connector-python